Divine Dreams /Emmy Lou Inc.
Hello I'm human nickname is Christopher L. Pennington I will turn 47 on 8-12-22. I am currently working my way back into the living. On Aug. 19 2007 I was the passenger in an alcohol related traffic accident that left me with a broken neck in 3 places a broken back in 8 places a crushed skull that lead to a 10 day coma. So here I am 3 back surgeries countless hours of physical and mental therapy, alcohol addiction, pharmaceutical  as well as street drug addiction. 2 suicide attempts and 2 unintentional overdoses and 14 years later. A changed man. A thankful man. A blessed man. This is a new year for me and for our species. The dark cloud has lifted off humanity. Its a bright and beautiful planet  our Father has created and so graciously allows us to inhibit. I believe we as a species is on the comeback. We have prayed and God has answered. Our fellow humans have started putting their Faith back to the Lord and themselves. I look forward to a very prosperous year financially and most of all Spiritually.